Move common code into shared folder
