+++ common/Perf/Azure.Test.Perf/Azure.Test.Perf.csproj
@@ -8,4 +8,8 @@
     <Reference Include="System.Net.Http" Condition="'$(TargetFrameworkIdentifier)' == '.NETFramework'" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Include="$(PerfStressSharedSources)PerfStressUtilities.cs" Link="Shared\%(RecursiveDir)\%(Filename)%(Extension)" />
+  </ItemGroup>
+
 </Project>
