+++ common/Perf/Azure.Test.Perf/PerfProgram.cs
@@ -1,13 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Azure.Test.PerfStress;
 using CommandLine;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
-using System.Reflection.Emit;
 using System.Runtime;
 using System.Text.Json;
 using System.Threading;
@@ -31,12 +31,12 @@ namespace Azure.Test.Perf
 
             if (testTypes.Any())
             {
-                var optionTypes = GetOptionTypes(testTypes);
-                await Parser.Default.ParseArguments(args, optionTypes).MapResult<PerfOptions, Task>(
+                var optionTypes = PerfStressUtilities.GetOptionTypes(testTypes);
+                await PerfStressUtilities.Parser.ParseArguments(args, optionTypes).MapResult<PerfOptions, Task>(
                     async o =>
                     {
                         var verbName = o.GetType().GetCustomAttribute<VerbAttribute>().Name;
-                        var testType = testTypes.Where(t => GetVerbName(t.Name) == verbName).Single();
+                        var testType = testTypes.Where(t => PerfStressUtilities.GetVerbName(t.Name) == verbName).Single();
                         await Run(testType, o);
                     },
                     errors => Task.CompletedTask
@@ -83,7 +83,7 @@ namespace Azure.Test.Perf
             Console.WriteLine();
 
             using var setupStatusCts = new CancellationTokenSource();
-            var setupStatusThread = PrintStatus("=== Setup ===", () => ".", newLine: false, setupStatusCts.Token);
+            var setupStatusThread = PerfStressUtilities.PrintStatus("=== Setup ===", () => ".", newLine: false, setupStatusCts.Token);
 
             using var cleanupStatusCts = new CancellationTokenSource();
             Thread cleanupStatusThread = null;
@@ -109,7 +109,8 @@ namespace Azure.Test.Perf
 
                         if (options.Warmup > 0)
                         {
-                            await RunTestsAsync(tests, options.Sync, options.Parallel, options.Rate, options.Warmup, "Warmup");
+                            await RunTestsAsync(tests, options.Sync, options.Parallel, options.Rate, options.Warmup, options.StatusInterval,
+                                "Warmup");
                         }
 
                         for (var i = 0; i < options.Iterations; i++)
@@ -119,7 +120,8 @@ namespace Azure.Test.Perf
                             {
                                 title += " " + (i + 1);
                             }
-                            await RunTestsAsync(tests, options.Sync, options.Parallel, options.Rate, options.Duration, title, options.JobStatistics, options.Latency);
+                            await RunTestsAsync(tests, options.Sync, options.Parallel, options.Rate, options.Duration, options.StatusInterval,
+                                title, options.JobStatistics, options.Latency);
                         }
                     }
                     finally
@@ -128,7 +130,7 @@ namespace Azure.Test.Perf
                         {
                             if (cleanupStatusThread == null)
                             {
-                                cleanupStatusThread = PrintStatus("=== Cleanup ===", () => ".", newLine: false, cleanupStatusCts.Token);
+                                cleanupStatusThread = PerfStressUtilities.PrintStatus("=== Cleanup ===", () => ".", newLine: false, cleanupStatusCts.Token);
                             }
 
                             await Task.WhenAll(tests.Select(t => t.CleanupAsync()));
@@ -141,7 +143,7 @@ namespace Azure.Test.Perf
                     {
                         if (cleanupStatusThread == null)
                         {
-                            cleanupStatusThread = PrintStatus("=== Cleanup ===", () => ".", newLine: false, cleanupStatusCts.Token);
+                            cleanupStatusThread = PerfStressUtilities.PrintStatus("=== Cleanup ===", () => ".", newLine: false, cleanupStatusCts.Token);
                         }
 
                         await tests[0].GlobalCleanupAsync();
@@ -161,7 +163,7 @@ namespace Azure.Test.Perf
         }
 
         private static async Task RunTestsAsync(IPerfTest[] tests, bool sync, int parallel, int? rate,
-            int durationSeconds, string title, bool jobStatistics = false, bool latency = false)
+            int durationSeconds, int statusIntervalSeconds, string title, bool jobStatistics = false, bool latency = false)
         {
             _completedOperations = new int[parallel];
             _lastCompletionTimes = new TimeSpan[parallel];
@@ -191,7 +193,7 @@ namespace Azure.Test.Perf
             var lastCompleted = 0;
 
             using var progressStatusCts = new CancellationTokenSource();
-            var progressStatusThread = PrintStatus(
+            var progressStatusThread = PerfStressUtilities.PrintStatus(
                 $"=== {title} ===" + Environment.NewLine +
                 "Current\t\tTotal",
                 () =>
@@ -202,7 +204,9 @@ namespace Azure.Test.Perf
                     return currentCompleted + "\t\t" + totalCompleted;
                 },
                 newLine: true,
-                progressStatusCts.Token);
+                progressStatusCts.Token,
+                statusIntervalSeconds
+                );
 
             Thread pendingOperationsThread = null;
             if (rate.HasValue)
@@ -375,7 +379,7 @@ namespace Azure.Test.Perf
             catch (Exception e)
             {
                 // Ignore if any part of the exception chain is type OperationCanceledException
-                if (!ContainsOperationCanceledException(e))
+                if (!PerfStressUtilities.ContainsOperationCanceledException(e))
                 {
                     throw;
                 }
@@ -405,113 +409,6 @@ namespace Azure.Test.Perf
             thread.Start();
 
             return thread;
-        }
-
-        // Run in dedicated thread instead of using async/await in ThreadPool, to ensure this thread has priority
-        // and never fails to run to due ThreadPool starvation.
-        private static Thread PrintStatus(string header, Func<object> status, bool newLine, CancellationToken token)
-        {
-            var thread = new Thread(() =>
-            {
-                Console.WriteLine(header);
-
-                bool needsExtraNewline = false;
-
-                while (!token.IsCancellationRequested)
-                {
-                    try
-                    {
-                        Sleep(TimeSpan.FromSeconds(1), token);
-                    }
-                    catch (OperationCanceledException)
-                    {
-                    }
-
-                    var obj = status();
-
-                    if (newLine)
-                    {
-                        Console.WriteLine(obj);
-                    }
-                    else
-                    {
-                        Console.Write(obj);
-                        needsExtraNewline = true;
-                    }
-                }
-
-                if (needsExtraNewline)
-                {
-                    Console.WriteLine();
-                }
-
-                Console.WriteLine();
-            });
-
-            thread.Start();
-
-            return thread;
-        }
-
-        private static void Sleep(TimeSpan timeout, CancellationToken token)
-        {
-            var sw = Stopwatch.StartNew();
-            while (sw.Elapsed < timeout)
-            {
-                if (token.IsCancellationRequested)
-                {
-                    // Simulate behavior of Task.Delay(TimeSpan, CancellationToken)
-                    throw new OperationCanceledException();
-                }
-
-                Thread.Sleep(TimeSpan.FromMilliseconds(10));
-            }
-        }
-
-        private static bool ContainsOperationCanceledException(Exception e)
-        {
-            if (e is OperationCanceledException)
-            {
-                return true;
-            }
-            else if (e.InnerException != null)
-            {
-                return ContainsOperationCanceledException(e.InnerException);
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        // Dynamically create option types with a "Verb" attribute
-        private static Type[] GetOptionTypes(IEnumerable<Type> testTypes)
-        {
-            var optionTypes = new List<Type>();
-
-            var ab = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("Options"), AssemblyBuilderAccess.Run);
-            var mb = ab.DefineDynamicModule("Options");
-
-            foreach (var t in testTypes)
-            {
-                var baseOptionsType = t.GetConstructors().First().GetParameters()[0].ParameterType;
-                var tb = mb.DefineType(t.Name + "Options", TypeAttributes.Public, baseOptionsType);
-
-                var attrCtor = typeof(VerbAttribute).GetConstructor(new Type[] { typeof(string), typeof(bool) });
-                var verbName = GetVerbName(t.Name);
-                tb.SetCustomAttribute(new CustomAttributeBuilder(attrCtor,
-                    new object[] { verbName, attrCtor.GetParameters()[1].DefaultValue }));
-
-                optionTypes.Add(tb.CreateType());
-            }
-
-            return optionTypes.ToArray();
-        }
-
-        private static string GetVerbName(string testName)
-        {
-            var lower = testName.ToLowerInvariant();
-            return lower.EndsWith("test") ? lower.Substring(0, lower.Length - 4) : lower;
         }
     }
 }
