+++ common/PerfStressShared/PerfStressUtilities.cs
@@ -1,0 +1,133 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using CommandLine;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Reflection;
+using System.Reflection.Emit;
+using System.Text.RegularExpressions;
+using System.Threading;
+
+namespace Azure.Test.PerfStress
+{
+    internal static class PerfStressUtilities
+    {
+        private static readonly Lazy<Parser> _defaultParser = new Lazy<Parser>(() => new Parser(settings =>
+        {
+            settings.CaseSensitive = false;
+            settings.HelpWriter = Console.Error;
+        }));
+
+        internal static Parser Parser => _defaultParser.Value;
+
+        internal static bool ContainsOperationCanceledException(Exception e)
+        {
+            if (e is OperationCanceledException)
+            {
+                return true;
+            }
+            else if (e.InnerException != null)
+            {
+                return ContainsOperationCanceledException(e.InnerException);
+            }
+            else
+            {
+                return false;
+            }
+        }
+
+        // Dynamically create option types with a "Verb" attribute
+        internal static Type[] GetOptionTypes(IEnumerable<Type> testTypes)
+        {
+            var optionTypes = new List<Type>();
+
+            var ab = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("Options"), AssemblyBuilderAccess.Run);
+            var mb = ab.DefineDynamicModule("Options");
+
+            foreach (var t in testTypes)
+            {
+                var baseOptionsType = t.GetConstructors().First().GetParameters()[0].ParameterType;
+                var tb = mb.DefineType(t.Name + "Options", TypeAttributes.Public, baseOptionsType);
+
+                var attrCtor = typeof(VerbAttribute).GetConstructor(new Type[] { typeof(string), typeof(bool) });
+                var verbName = GetVerbName(t.Name);
+                tb.SetCustomAttribute(new CustomAttributeBuilder(attrCtor,
+                    new object[] { verbName, attrCtor.GetParameters()[1].DefaultValue }));
+
+                optionTypes.Add(tb.CreateType());
+            }
+
+            return optionTypes.ToArray();
+        }
+
+        internal static string GetVerbName(string testName)
+        {
+            // Remove suffix "Test" if exists
+            return Regex.Replace(testName, "Test$", string.Empty, RegexOptions.IgnoreCase);
+        }
+
+        // Run in dedicated thread instead of using async/await in ThreadPool, to ensure this thread has priority
+        // and never fails to run to due ThreadPool starvation.
+        internal static Thread PrintStatus(string header, Func<object> status, bool newLine, CancellationToken token, int intervalSeconds = 1)
+        {
+            var thread = new Thread(() =>
+            {
+                Console.WriteLine(header);
+
+                bool needsExtraNewline = false;
+
+                while (!token.IsCancellationRequested)
+                {
+                    try
+                    {
+                        Sleep(TimeSpan.FromSeconds(intervalSeconds), token);
+                    }
+                    catch (OperationCanceledException)
+                    {
+                    }
+
+                    var obj = status();
+
+                    if (newLine)
+                    {
+                        Console.WriteLine(obj);
+                    }
+                    else
+                    {
+                        Console.Write(obj);
+                        needsExtraNewline = true;
+                    }
+                }
+
+                if (needsExtraNewline)
+                {
+                    Console.WriteLine();
+                }
+
+                Console.WriteLine();
+            });
+
+            thread.Start();
+
+            return thread;
+        }
+
+        internal static void Sleep(TimeSpan timeout, CancellationToken token)
+        {
+            var sw = Stopwatch.StartNew();
+            while (sw.Elapsed < timeout)
+            {
+                if (token.IsCancellationRequested)
+                {
+                    // Simulate behavior of Task.Delay(TimeSpan, CancellationToken)
+                    throw new OperationCanceledException();
+                }
+
+                Thread.Sleep(TimeSpan.FromMilliseconds(10));
+            }
+        }
+    }
+}
