+++ common/Stress/Azure.Test.Stress/Azure.Test.Stress.csproj
@@ -6,4 +6,8 @@
     <PackageReference Include="System.Reflection.Emit" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Include="$(PerfStressSharedSources)PerfStressUtilities.cs" Link="Shared\%(RecursiveDir)\%(Filename)%(Extension)" />
+  </ItemGroup>
+
 </Project>
