+++ eng/Directory.Build.Data.props
@@ -159,6 +159,10 @@
     <ServerGarbageCollection>true</ServerGarbageCollection>
   </PropertyGroup>
 
+  <PropertyGroup Condition="'$(IsPerfProject)' == 'true' or '$(IsStressProject)' == 'true'">
+    <PerfStressSharedSources>$(MSBuildThisFileDirectory)/../common/PerfStressShared/</PerfStressSharedSources>
+  </PropertyGroup>
+
   <!-- Code generation settings -->
   <PropertyGroup>
     <!-- Don't use code generator bundled Azure.Core shared source for code generation -->
